+++ fdbbackup/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(FDBBACKUP_SRCS
 	BackupTLSConfig.cpp
+	Decode.cpp
 	backup.actor.cpp)
 
 add_flow_target(EXECUTABLE NAME fdbbackup SRCS ${FDBBACKUP_SRCS})
@@ -13,6 +14,7 @@ target_include_directories(fdbconvert PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/inclu
 target_link_libraries(fdbconvert PRIVATE fdbclient)
 
 set(FDBDECODE_SRCS
+  Decode.cpp
 	BackupTLSConfig.cpp
 	FileDecoder.actor.cpp)
 add_flow_target(EXECUTABLE NAME fdbdecode SRCS ${FDBDECODE_SRCS})
