+++ fdbbackup/FileDecoder.actor.cpp
@@ -34,6 +34,7 @@
 #include "fdbbackup/BackupTLSConfig.h"
 #include "fdbclient/BuildFlags.h"
 #include "fdbbackup/FileConverter.h"
+#include "fdbbackup/Decode.h"
 #include "fdbclient/BackupAgent.actor.h"
 #include "fdbclient/BackupContainer.h"
 #include "fdbclient/BackupContainerFileSystem.h"
@@ -246,61 +247,6 @@ struct DecodeParams : public ReferenceCounted<DecodeParams> {
 		IKnobCollection::getMutableGlobalKnobCollection().initialize(Randomize::False, IsSimulated::False);
 	}
 };
-
-// Decode an ASCII string, e.g., "\x15\x1b\x19\x04\xaf\x0c\x28\x0a",
-// into the binary string. Set "err" to true if the format is invalid.
-// Note ',' '\' '," ';' are escaped by '\'. Normal characters can be
-// unencoded into HEX, but not recommended.
-std::string decode_hex_string(std::string line, bool& err) {
-	size_t i = 0;
-	std::string ret;
-
-	while (i <= line.length()) {
-		switch (line[i]) {
-		case '\\':
-			if (i + 2 > line.length()) {
-				std::cerr << "Invalid hex string at: " << i << "\n";
-				err = true;
-				return ret;
-			}
-			switch (line[i + 1]) {
-				char ent, save;
-			case '"':
-			case '\\':
-			case ' ':
-			case ';':
-				line.erase(i, 1);
-				break;
-			case 'x':
-				if (i + 4 > line.length()) {
-					std::cerr << "Invalid hex string at: " << i << "\n";
-					err = true;
-					return ret;
-				}
-				char* pEnd;
-				save = line[i + 4];
-				line[i + 4] = 0;
-				ent = char(strtoul(line.data() + i + 2, &pEnd, 16));
-				if (*pEnd) {
-					std::cerr << "Invalid hex string at: " << i << "\n";
-					err = true;
-					return ret;
-				}
-				line[i + 4] = save;
-				line.replace(i, 4, 1, ent);
-				break;
-			default:
-				std::cerr << "Invalid hex string at: " << i << "\n";
-				err = true;
-				return ret;
-			}
-		default:
-			i++;
-		}
-	}
-
-	return line.substr(0, i);
-}
 
 // Parses and returns a ";" separated HEX encoded strings. So the ";" in
 // the string should be escaped as "\;".
