+++ fdbbackup/backup.actor.cpp
@@ -21,6 +21,7 @@
 #include "flow/ApiVersion.h"
 #include "fmt/format.h"
 #include "fdbbackup/BackupTLSConfig.h"
+#include "fdbbackup/Decode.h"
 #include "fdbclient/JsonBuilder.h"
 #include "flow/Arena.h"
 #include "flow/ArgParseUtil.h"
@@ -3065,61 +3066,6 @@ ACTOR Future<Void> modifyBackup(Database db, std::string tagName, BackupModifyOp
 	}
 
 	return Void();
-}
-
-// Decode an ASCII string, e.g., "\x15\x1b\x19\x04\xaf\x0c\x28\x0a",
-// into the binary string. Set "err" to true if the format is invalid.
-// Note ',' '\' '," ';' are escaped by '\'. Normal characters can be
-// unencoded into HEX, but not recommended.
-std::string decode_hex_string(std::string line, bool& err) {
-	size_t i = 0;
-	std::string ret;
-
-	while (i <= line.length()) {
-		switch (line[i]) {
-		case '\\':
-			if (i + 2 > line.length()) {
-				std::cerr << "Invalid hex string at: " << i << "\n";
-				err = true;
-				return ret;
-			}
-			switch (line[i + 1]) {
-				char ent, save;
-			case '"':
-			case '\\':
-			case ' ':
-			case ';':
-				line.erase(i, 1);
-				break;
-			case 'x':
-				if (i + 4 > line.length()) {
-					std::cerr << "Invalid hex string at: " << i << "\n";
-					err = true;
-					return ret;
-				}
-				char* pEnd;
-				save = line[i + 4];
-				line[i + 4] = 0;
-				ent = char(strtoul(line.data() + i + 2, &pEnd, 16));
-				if (*pEnd) {
-					std::cerr << "Invalid hex string at: " << i << "\n";
-					err = true;
-					return ret;
-				}
-				line[i + 4] = save;
-				line.replace(i, 4, 1, ent);
-				break;
-			default:
-				std::cerr << "Invalid hex string at: " << i << "\n";
-				err = true;
-				return ret;
-			}
-		default:
-			i++;
-		}
-	}
-
-	return line.substr(0, i);
 }
 
 static std::vector<std::vector<StringRef>> parseLine(std::string& line, bool& err, bool& partial) {
